refactor(cli): separate parseStoreConfig from loadStoreConfig
