+++ packages/cli/src/config/index.ts
@@ -1,4 +1,5 @@
 export * from "./commonSchemas.js";
 export * from "./loadConfig.js";
 export * from "./loadStoreConfig.js";
+export * from "./parseStoreConfig.js";
 export * from "./validation.js";
