+++ packages/cli/src/config/parseStoreConfig.test-d.ts
@@ -1,8 +1,8 @@
 import { describe, expectTypeOf } from "vitest";
 import { z } from "zod";
-import { StoreConfig, StoreUserConfig } from "./loadStoreConfig.js";
+import { StoreConfig, StoreUserConfig } from "./parseStoreConfig.js";
 
-describe("loadStoreConfig", () => {
+describe("StoreUserConfig", () => {
   // Typecheck manual interfaces against zod
   expectTypeOf<StoreUserConfig>().toEqualTypeOf<z.input<typeof StoreConfig>>();
   // type equality isn't deep for optionals
