+++ packages/cli/src/index.ts
@@ -1,7 +1,8 @@
-import { loadStoreConfig, parseStoreConfig } from "./config/loadStoreConfig.js";
+import { parseStoreConfig } from "./config/parseStoreConfig.js";
+import { loadStoreConfig } from "./config/loadStoreConfig.js";
 import { renderTablesFromConfig } from "./render-table/renderTablesFromConfig.js";
 import { renderTable } from "./render-table/renderTable.js";
 
-export type { StoreUserConfig, StoreConfig } from "./config/loadStoreConfig.js";
+export type { StoreUserConfig, StoreConfig } from "./config/parseStoreConfig.js";
 
 export { loadStoreConfig, parseStoreConfig, renderTablesFromConfig, renderTable };
